updated upload logic and fixed orphans and email

+++ users/models.py
@@ -4,7 +4,7 @@ from django.db.models.deletion import CASCADE
 from django.contrib.auth.models import User
 from django.db.models.fields import DecimalField, related
 from phonenumber_field.modelfields import PhoneNumberField
-from referrals.models import Referral, OrphanList
+from referrals.models import Referral, OrphanList, orphanReceipt, receipts
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
@@ -17,6 +17,7 @@ class Profile(models.Model):
     wallet= models.DecimalField(max_digits=6, decimal_places=2, default=0)
     num_of_refers = models.IntegerField(default=0)
     links_created = models.IntegerField(default=0)
+    signup_refs = models.IntegerField(default=0)
     code = models.CharField(max_length=30, blank= True)
     recommended_by = models.ForeignKey(User, on_delete=CASCADE, blank=True, related_name="ref_by", null=True)
 
@@ -26,15 +27,15 @@ class Profile(models.Model):
 @receiver(post_save, sender = Profile)
 def post_save_profile(sender, instance, created, *args, **kwargs):
     if created:
-        #check orphans
+        #check orphans for partner orders
         if OrphanList.objects.filter(refereeEmail = instance.user.email).exists():
             # update the username in the ref object
-            ref_obj = OrphanList.objects.filter(refereeEmail = instance.user.email)[0].referral_obj
+            ref_obj = OrphanList.objects.get(refereeEmail = instance.user.email).referral_obj
             ref_obj.referee_username = instance.user.username
             ref_obj.referee_has_account = True
             
             #update the wallet and num refers of referee
-            referee = User.objects.filter(username = ref_obj.referee_username)[0]
+            referee = User.objects.get(username = ref_obj.referee_username)
             referee.profile.wallet = referee.profile.wallet + ref_obj.referee_cashback
             referee.profile.num_of_refers = referee.profile.num_of_refers + 1
             referee.profile.save()
@@ -42,5 +43,21 @@ def post_save_profile(sender, instance, created, *args, **kwargs):
             ref_obj.save()
          
             #delete the orphan list object
-            obj = OrphanList.objects.filter(refereeEmail = instance.user.email)[0]
+            obj = OrphanList.objects.get(refereeEmail = instance.user.email)
+            obj.delete()
+        
+        #check orphan receipts for amazon order
+        if orphanReceipt.objects.filter(referee = instance.Phone_Number).exists():
+            #update username of referee on receipt object
+            receipt_obj = orphanReceipt.objects.get(referee = str(instance.Phone_Number)).referral_obj
+            receipt_obj.referee = instance.user.username
+            print(receipt_obj.referee )
+            receipt_obj.save()
+
+            #update the number of refers of referee
+            referee = User.objects.get(username = receipt_obj.referer)
+            referee.profile.num_of_refers = referee.profile.num_of_refers + 1
+            referee.profile.save()
+
+            obj = orphanReceipt.objects.get(referee = str(instance.Phone_Number))
             obj.delete()
